# Server/app/api/projects.py
"""
Projects API endpoints
"""

from fastapi import APIRouter, Depends, HTTPException, status
from pydantic import BaseModel
from typing import List, Optional
import structlog

from app.core.exceptions import NotFoundError, ValidationError
from app.services.auth_service import get_current_active_user
from app.models.user import User

logger = structlog.get_logger()

router = APIRouter()

# Pydantic models
class ProjectResponse(BaseModel):
    success: bool
    message: str
    data: Optional[dict] = None

class ProjectCreate(BaseModel):
    name: str
    description: Optional[str] = None

class ProjectUpdate(BaseModel):
    name: Optional[str] = None
    description: Optional[str] = None

@router.get("/")
async def list_projects(current_user: User = Depends(get_current_active_user)):
    """List all projects"""
    logger.info("Listing projects", user=current_user.username)
    return {"message": "Projects endpoint - implementation pending"}

@router.get("/{item_id}")
async def get_project(item_id: str, current_user: User = Depends(get_current_active_user)):
    """Get project by ID"""
    logger.info(f"Getting project {item_id}", user=current_user.username)
    raise NotFoundError(f"Project {item_id} not found")

@router.post("/")
async def create_project(data: ProjectCreate, current_user: User = Depends(get_current_active_user)):
    """Create new project"""
    logger.info("Creating project", requester=current_user.username, name=data.name)
    return ProjectResponse(success=True, message="Project created successfully")

@router.put("/{item_id}")
async def update_project(
    item_id: str,
    data: ProjectUpdate,
    current_user: User = Depends(get_current_active_user)
):
    """Update project"""
    logger.info("Updating project", item_id=item_id, requester=current_user.username)
    return ProjectResponse(success=True, message="Project updated successfully")

@router.delete("/{item_id}")
async def delete_project(item_id: str, current_user: User = Depends(get_current_active_user)):
    """Delete project"""
    logger.info("Deleting project", item_id=item_id, requester=current_user.username)
    return ProjectResponse(success=True, message="Project deleted successfully")