# Server/app/api/scripts.py
"""
Test Scripts API endpoints
"""

from fastapi import APIRouter, Depends, HTTPException, status
from pydantic import BaseModel
from typing import List, Optional
import structlog

from app.core.exceptions import NotFoundError, ValidationError
from app.services.auth_service import get_current_active_user
from app.models.user import User

logger = structlog.get_logger()

router = APIRouter()

# Pydantic models
class ScriptResponse(BaseModel):
    success: bool
    message: str
    data: Optional[dict] = None

class ScriptCreate(BaseModel):
    name: str
    description: Optional[str] = None
    content: dict

class ScriptUpdate(BaseModel):
    name: Optional[str] = None
    description: Optional[str] = None
    content: Optional[dict] = None

@router.get("/")
async def list_scripts(current_user: User = Depends(get_current_active_user)):
    """List all test scripts"""
    logger.info("Listing scripts", user=current_user.username)
    return {"message": "Scripts endpoint - implementation pending"}

@router.get("/{item_id}")
async def get_script(item_id: str, current_user: User = Depends(get_current_active_user)):
    """Get test script by ID"""
    logger.info(f"Getting script {item_id}", user=current_user.username)
    raise NotFoundError(f"Script {item_id} not found")

@router.post("/")
async def create_script(data: ScriptCreate, current_user: User = Depends(get_current_active_user)):
    """Create new test script"""
    logger.info("Creating script", requester=current_user.username, name=data.name)
    return ScriptResponse(success=True, message="Script created successfully")

@router.put("/{item_id}")
async def update_script(
    item_id: str,
    data: ScriptUpdate,
    current_user: User = Depends(get_current_active_user)
):
    """Update test script"""
    logger.info("Updating script", item_id=item_id, requester=current_user.username)
    return ScriptResponse(success=True, message="Script updated successfully")

@router.delete("/{item_id}")
async def delete_script(item_id: str, current_user: User = Depends(get_current_active_user)):
    """Delete test script"""
    logger.info("Deleting script", item_id=item_id, requester=current_user.username)
    return ScriptResponse(success=True, message="Script deleted successfully")